# pkg/api_client/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API register API v1",
    "description": "API van het API register (apis.developer.overheid.nl)",
    "version": "1.0.0",
    "contact": {
      "name": "Team developer.overheid.nl",
      "email": "developer.overheid@geonovum.nl",
      "url": "https://apis.developer.overheid.nl"
    }
  },
  "servers": [
    {
      "url": "https://api.developer.overheid.nl/apis/v1"
    }
  ],
  "tags": [
    {
      "name": "API's",
      "description": "Alle API operaties die horen bij API's."
    }
  ],
  "paths": {
    "/apis": {
      "get": {
        "operationId": "listApis",
        "description": "Alle API's ophalen",
        "summary": "Alle API's ophalen",
        "tags": [
          "API's"
        ],
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "https://static.developer.overheid.nl/adr/components.yaml#/headers/API-Version"
              },
              "Link": {
                "$ref": "https://static.developer.overheid.nl/adr/components.yaml#/headers/Link"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Api"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createApiFromOas",
        "summary": "Registreer een nieuwe API met een OpenAPI URL",
        "description": "Valideer en sla een nieuwe API op via een OpenAPI-bestand (OAS) dat via een URL wordt aangeleverd.",
        "tags": [
          "API's"
        ],
        "requestBody": {
          "description": "Informatie over het te registreren OAS-bestand",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "oasUrl"
                ],
                "properties": {
                  "oasUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Publiek toegankelijke URL naar een OpenAPI 3.x-bestand"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "De API is succesvol geregistreerd",
            "headers": {
              "API-Version": {
                "$ref": "https://static.developer.overheid.nl/adr/components.yaml#/headers/API-Version"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Api"
                }
              }
            }
          },
          "400": {
            "description": "Fout in aanvraag of ongeldige OAS URL"
          },
          "422": {
            "description": "OAS-bestand kon niet gevalideerd worden"
          }
        }
      }
    },
    "/apis/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/id"
        }
      ],
      "get": {
        "operationId": "retrieveApi",
        "description": "API ophalen",
        "summary": "API ophalen",
        "tags": [
          "API's"
        ],
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "https://static.developer.overheid.nl/adr/components.yaml#/headers/API-Version"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Api"
                }
              }
            }
          },
          "404": {
            "$ref": "https://static.developer.overheid.nl/adr/components.yaml#/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Api": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "oasUri": {
            "type": "string"
          },
          "docsUri": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          },
          "adrScore": {
            "type": "string"
          },
          "repositoryUri": {
            "type": "string"
          },
          "organisation": {
            "type": "object",
            "properties": {
              "label": {
                "type": "string"
              },
              "uri": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "description": "id",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    }
  }
}